testing

+++ Components/Internal/TrackMotor.py
@@ -13,7 +13,7 @@ class TrackMotor:
 
         byte_arr.append(int(abs(ControllerData.joystick1[0]) * 255))
         byte_arr.append(bool(ControllerData.joystick1[0] > 0))
-        byte_arr.append(abs(ControllerData.joystick1[1]) * 255)
+        byte_arr.append(int(abs(ControllerData.joystick1[1]) * 255))
         byte_arr.append(bool(ControllerData.joystick1[1] > 0))
         print(len(byte_arr))
